requirements-dev.txt
-r requirements.txt

pre-commit==3.6.0
ruff==0.4.2
pytest==7.4.4
pytest-django==4.7.0
pytest-mock==3.12.0
schema==0.7.5
python-dotenv==1.0.1